feat: Add Avatar component for displaying user avatars
The commit adds a new Avatar component to the app, which allows for the display of user avatars. This component utilizes the "@radix-ui/react-avatar" package to provide a customizable and visually appealing avatar. The addition of this component enhances the user experience by personalizing user profiles and improving the overall visual design of the app.