feat: Update useCoverImage hook to reset URL when opening
This commit updates the useCoverImage hook in the hooks/use-cover-image.tsx file. The onOpen function is modified to reset the URL to undefined when opening the cover image modal. This change ensures that the previous URL is cleared when selecting a new cover image, improving the functionality and user experience of the application.